+++ src/components/App.tsx
@@ -1,10 +1,10 @@
-import Header from "./Layout/Header";
+import Header from "./Layout/Header/Header";
 import React from "react";
 import { Outlet } from "react-router-dom";
 import "../styles/reset.css";
 import "../styles/common.css";
 import "../styles/global.css";
-import Footer from "../pages/Footer/Footer";
+import Footer from "./Layout/Footer/Footer";
 
 function App() {
   return (
