+++ src/components/Layout/Footer/FaqPage.tsx
