+++ src/components/Layout/Footer/PrivacyPage.tsx
