+++ src/components/Layout/UI/Input.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Input() {
+  return <></>;
+}
+
+export default Input;
