+++ src/pages/LoginPage/LoginPage.scss
@@ -1,7 +1,6 @@
 .login-page-main {
   width: 640px;
   margin: 0 auto;
-  padding-top: 60px;
 }
 
 .form {
@@ -12,46 +11,52 @@
 .go-home-logo {
   display: block;
   text-align: center;
-  margin-top: 60px;
   margin-bottom: 40px;
-}
-
-label {
-  display: inline-block;
-  color: #1f2937;
-  font-size: var(--fontsize-18); /*18px;*/
-  font-weight: 700;
-  margin-bottom: 16px;
-}
-
-input {
-  width: 100%;
-  height: 56px;
-  background-color: #f3f4f6;
-  padding: 0px 24px;
-  border: none;
-  border-radius: 12px;
-  font-size: 1em;
-  line-height: 24px;
-  font-family: inherit;
-  font-weight: 400;
 }
 
 .input-area {
   margin: 24px 0;
   width: 100%;
-}
 
-.input-area input {
-  width: 100%;
-}
+  label {
+    display: inline-block;
+    color: #1f2937;
+    font-size: var(--fontsize-18); /*18px;*/
+    font-weight: 700;
+    margin-bottom: 16px;
+  }
 
-input::placeholder {
-  color: #9ca3af;
-}
+  input {
+    width: 100%;
+    height: 56px;
+    background-color: #f3f4f6;
+    padding: 0px 24px;
+    border: none;
+    border-radius: 12px;
+    font-size: 1em;
+    line-height: 24px;
+    font-family: inherit;
+    font-weight: 400;
+  }
 
-input:focus {
-  outline-color: var(--blue);
+  input::placeholder {
+    color: #9ca3af;
+  }
+
+  input:focus {
+    outline-color: var(--blue);
+  }
+
+  h3 {
+    color: #1f2937;
+    font-size: 1em;
+    line-height: 24px;
+    font-weight: 500;
+  }
+
+  .password-input {
+    position: relative;
+  }
 }
 
 .acc-button {
@@ -75,13 +80,6 @@ input:focus {
   align-items: center;
 }
 
-h3 {
-  color: #1f2937;
-  font-size: 1em;
-  line-height: 24px;
-  font-weight: 500;
-}
-
 .easy-login img {
   margin-left: 16px;
 }
@@ -93,19 +91,11 @@ h3 {
   font-weight: 500;
   text-align: center;
   margin-bottom: 60px;
-}
 
-.signup a {
-  color: var(--hyperlink-blue);
-  text-decoration: underline;
-}
-
-input img {
-  inset: 0;
-}
-
-.password-input {
-  position: relative;
+  a {
+    color: var(--hyperlink-blue);
+    text-decoration: underline;
+  }
 }
 
 .visibility-button {
@@ -116,11 +106,6 @@ input img {
   width: auto;
 }
 
-/* 비밀번호 가시성 토글 기본 비가시성 */
-.visibility-on {
-  display: none;
-}
-
 /* 유효성 에러 text 기본 비가시성 */
 .error-message {
   margin: 8px 16px;
@@ -128,18 +113,6 @@ input img {
   font-weight: 600;
   font-size: var(--fontsize-15);
 
-  display: none;
-}
-
-/* 우선 순위 때문에 코드 맨 아래에 위치해야함 */
-
-/* 가시성 설정 */
-.visible-maker {
-  display: inline-block;
-}
-
-/* 비가시성 설정 */
-.hidden-maker {
   display: none;
 }
 
@@ -161,8 +134,8 @@ input img {
 /* Mobile: 767px 이하 */
 @media screen and (max-width: 767px) {
   /*기본 패딩과 최대 넓이*/
-  main {
-    padding: 0 16px;
+  .login-page-main {
+    padding: 60px 16px 0;
     max-width: 400px;
   }
 
@@ -170,17 +143,15 @@ input img {
   .go-home-logo {
     margin: 24px 0;
   }
+
   .go-home-logo img {
     width: 198px;
   }
 
-  label {
-    font-size: 0.875em; /*14px;*/
-  }
-
-  input {
-    margin-top: 8px;
-    margin-bottom: 16px;
+  .input-area {
+    label {
+      font-size: 0.875em; /*14px;*/
+    }
   }
 
   .pill-button {
