+++ src/pages/SignupPage/SignupPage.tsx
@@ -1,18 +1,14 @@
 import React from "react";
 import "./SignupPage.css";
-import logoImg from "../../assets/images/icons/panda-market-logo.svg";
-import EasyLogin from "../../UI/EasyLogin";
+import EasyLogin from "../../components/Layout/UI/EasyLogin";
 import { Link } from "react-router-dom";
-import TogglePassword from "../../UI/togglePassword";
+import TogglePassword from "../../components/Layout/UI/TogglePassword";
+import GoHomeLogo from "../../components/Layout/UI/GoHomeLogo";
 
 function SignupPage() {
   return (
     <div className="signup-page-main">
-      <section className="go-home-logo">
-        <a href="/">
-          <img src={logoImg} alt="판다 마켓 홈" width="396" />
-        </a>
-      </section>
+      <GoHomeLogo />
       <section className="input">
         <form method="post" id="signup" className="form">
           <div className="input-area">
